working on test logic

+++ vme-service/src/main/java/org/fao/fi/vme/dao/VmeDao.java
@@ -24,7 +24,7 @@ public class VmeDao extends Dao {
 		return this.generateTypedQuery(em, clazz).getResultList();
 	}
 
-	public Vme findVme(int id) {
+	public Vme findVme(Long id) {
 		return em.find(Vme.class, id);
 	}
